requirements.txt
# gcp-security-agent requirements

# GCP SDKs
google-auth>=2.28.0
google-cloud-run>=0.10.0
google-cloud-logging>=3.9.0
google-cloud-resource-manager>=1.12.0

# Vertex AI / Gemini
google-genai>=0.8.0

# Templating (remediation commands)
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
